requirements.txt
# Requirements

Flask==0.12.2
Flask-API==0.7.1
Jinja2==2.9.6
requests==2.19.1
MySQL-connector-python==8.0.12